Fix a couple of MSVC compiler warnings;
FossilOrigin-Name: 748c9109c973e1d3e914adbe533f12af7783bfe4

+++ src/vdbetrace.c
@@ -85,7 +85,7 @@ char *sqlite3VdbeExpandSql(
       const char *zStart = zRawSql;
       while( *(zRawSql++)!='\n' && *zRawSql );
       sqlite3StrAccumAppend(&out, "-- ", 3);
-      sqlite3StrAccumAppend(&out, zStart, zRawSql-zStart);
+      sqlite3StrAccumAppend(&out, zStart, (int)(zRawSql-zStart));
     }
   }else{
     while( zRawSql[0] ){
